refactor(utils): split from usage
